Fix missing sequence nr and signature updates at refresh offers

+++ network/src/main/java/io/bitsquare/p2p/peers/getdata/messages/GetDataRequest.java
@@ -3,5 +3,5 @@ package io.bitsquare.p2p.peers.getdata.messages;
 import io.bitsquare.p2p.Message;
 
 public interface GetDataRequest extends Message {
-    long getNonce();
+    int getNonce();
 }
